requirements.txt
opencv-python>=4.8.0.76

numpy>=2.1.0
pillow